requirements.txt
# Tested with Python 3.9–3.10 on Linux x86_64
numpy==1.24.4
pandas==2.1.4
scikit-learn==1.3.2
torch==2.1.2           # install CUDA build if GPU is available; otherwise CPU wheel
tqdm==4.66.1
networkx==3.1
xgboost==1.7.6
rdkit-pypi==2023.3.1b1
